core: merge fuzzy_workflow module into heavy_tests
